+++ src/SpeedBonus.java
@@ -6,7 +6,7 @@ public class SpeedBonus extends Bonus {
 
     private final String IMG_DIR = "./MediaFiles/image11.png";
 
-    public SpeedBonus(int xLoc, int yLoc) {
+    public SpeedBonus(double xLoc, double yLoc) {
         super(xLoc, yLoc);
         super.setImage( new Image(Paths.get(IMG_DIR).toUri().toString()));
         super.initView();
